Split the test files back.
